+++ platform/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.1.1] platform
 ### Releases
 - [NPM v0.1.1] https://www.npmjs.com/package/@codat/platform/v/0.1.1 - platform
+
+## 2023-09-05 00:04:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] platform
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@codat/platform/v/0.2.0 - platform